+++ mobile/android/components/extensions/ext-tabs.js
@@ -8,6 +8,12 @@ ChromeUtils.defineModuleGetter(
   "resource://gre/modules/PromiseUtils.jsm"
 );
 
+ChromeUtils.defineModuleGetter(
+  this,
+  "GeckoViewTabBridge",
+  "resource://gre/modules/GeckoViewTab.jsm"
+);
+
 const getBrowserWindow = window => {
   return window.docShell.rootTreeItem.domWindow;
 };
@@ -339,7 +345,8 @@ this.tabs = class extends ExtensionAPI {
             nativeTab = BrowserApp.addTab(url, options);
           } else {
             options.extensionId = context.extension.id;
-            nativeTab = await BrowserApp.createNewTab(url, options);
+            options.url = url;
+            nativeTab = await GeckoViewTabBridge.createNewTab(options);
           }
 
           if (createProperties.url) {
@@ -352,6 +359,27 @@ this.tabs = class extends ExtensionAPI {
         async remove(tabs) {
           if (!Array.isArray(tabs)) {
             tabs = [tabs];
+          }
+
+          if (!Services.androidBridge.isFennec) {
+            await Promise.all(
+              tabs.map(async tabId => {
+                const windowId = GeckoViewTabBridge.tabIdToWindowId(tabId);
+                const window = windowTracker.getWindow(
+                  windowId,
+                  context,
+                  false
+                );
+                if (!window) {
+                  throw new ExtensionError(`Invalid tab ID ${tabId}`);
+                }
+                await GeckoViewTabBridge.closeTab({
+                  window,
+                  extensionId: context.extension.id,
+                });
+              })
+            );
+            return;
           }
 
           for (let tabId of tabs) {