+++ mobile/android/geckoview/src/androidTest/java/org/mozilla/geckoview/test/WebExtensionTest.kt
@@ -29,7 +29,9 @@ import java.util.UUID
 class WebExtensionTest : BaseSessionTest() {
     companion object {
         val TEST_ENDPOINT: String = "http://localhost:4243"
-        val TABS_BACKGROUND: String = "resource://android/assets/web_extensions/tabs/"
+        val TABS_CREATE_BACKGROUND: String = "resource://android/assets/web_extensions/tabs-create/"
+        val TABS_CREATE_REMOVE_BACKGROUND: String = "resource://android/assets/web_extensions/tabs-create-remove/"
+        val TABS_REMOVE_BACKGROUND: String = "resource://android/assets/web_extensions/tabs-remove/"
         val MESSAGING_BACKGROUND: String = "resource://android/assets/web_extensions/messaging/"
         val MESSAGING_CONTENT: String = "resource://android/assets/web_extensions/messaging-content/"
     }
@@ -120,6 +122,8 @@ class WebExtensionTest : BaseSessionTest() {
     // This test
     // - Listen for a new tab request from a web extension
     // - Registers a web extension
+    // - Waits for onNewTab request
+    // - Verify that request came from right extension
     @Test
     fun testBrowserTabsCreate() {
         val tabsCreateResult = GeckoResult<Void>()
@@ -134,10 +138,81 @@ class WebExtensionTest : BaseSessionTest() {
             }
         }
         sessionRule.runtime.webExtensionController.tabDelegate = tabDelegate
-        tabsExtension = WebExtension(TABS_BACKGROUND)
+        tabsExtension = WebExtension(TABS_CREATE_BACKGROUND)
 
         sessionRule.waitForResult(sessionRule.runtime.registerWebExtension(tabsExtension))
         sessionRule.waitForResult(tabsCreateResult)
+
+        sessionRule.waitForResult(sessionRule.runtime.unregisterWebExtension(tabsExtension))
+    }
+
+    // This test
+    // - Create and assign WebExtension TabDelegate to handle creation and closing of tabs
+    // - Registers a WebExtension
+    // - Extension requests creation of new tab
+    // - TabDelegate handles creation of new tab
+    // - Extension requests removal of newly created tab
+    // - TabDelegate handles closing of newly created tab
+    // - Verify that close request came from right extension and targeted session
+    @Test
+    fun testBrowserTabsCreateBrowserTabsRemove() {
+        val onCloseRequestResult = GeckoResult<Void>()
+        var tabsExtension : WebExtension? = null
+        var extensionCreatedSession : GeckoSession? = null
+
+        val tabDelegate = object : WebExtensionController.TabDelegate {
+            override fun onNewTab(source: WebExtension?, uri: String?): GeckoResult<GeckoSession> {
+                extensionCreatedSession = GeckoSession(sessionRule.session.settings)
+                return GeckoResult.fromValue(extensionCreatedSession)
+            }
+
+            override fun onCloseTab(source: WebExtension?, session: GeckoSession): GeckoResult<AllowOrDeny> {
+                Assert.assertEquals(tabsExtension, source)
+                Assert.assertNotEquals(null, extensionCreatedSession)
+                Assert.assertEquals(extensionCreatedSession, session)
+                onCloseRequestResult.complete(null)
+                return GeckoResult.ALLOW;
+            }
+        }
+
+        sessionRule.runtime.webExtensionController.tabDelegate = tabDelegate
+        tabsExtension = WebExtension(TABS_CREATE_REMOVE_BACKGROUND)
+
+        sessionRule.waitForResult(sessionRule.runtime.registerWebExtension(tabsExtension))
+        sessionRule.waitForResult(onCloseRequestResult)
+
+        sessionRule.waitForResult(sessionRule.runtime.unregisterWebExtension(tabsExtension))
+    }
+
+    // This test
+    // - Create and assign WebExtension TabDelegate to handle closing of tabs
+    // - Create new GeckoSession for WebExtension to close
+    // - Load url that will allow extension to identify the tab
+    // - Registers a WebExtension
+    // - Extension finds the tab by url and removes it
+    // - TabDelegate handles closing of the tab
+    // - Verify that request targets previously created GeckoSession
+    @Test
+    fun testBrowserTabsRemove() {
+        val onCloseRequestResult = GeckoResult<Void>()
+        val existingSession = sessionRule.createOpenSession()
+
+        val tabDelegate = object : WebExtensionController.TabDelegate {
+            override fun onCloseTab(source: WebExtension?, session: GeckoSession): GeckoResult<AllowOrDeny> {
+                Assert.assertEquals(existingSession, session)
+                onCloseRequestResult.complete(null)
+                return GeckoResult.ALLOW;
+            }
+        }
+
+        existingSession.loadTestPath("$HELLO_HTML_PATH?tabToClose")
+        existingSession.waitForPageStop()
+
+        sessionRule.runtime.webExtensionController.tabDelegate = tabDelegate
+        val tabsExtension = WebExtension(TABS_REMOVE_BACKGROUND)
+
+        sessionRule.waitForResult(sessionRule.runtime.registerWebExtension(tabsExtension))
+        sessionRule.waitForResult(onCloseRequestResult)
 
         sessionRule.waitForResult(sessionRule.runtime.unregisterWebExtension(tabsExtension))
     }